Revert "MTDSA-16151: Copy V1 Get Crystallisation to V2 (#106)"
This reverts commit 115ee9c7a85585d0097751652f53f5787afce2fc.
