--- app/v2/controllers/requestParsers/RetrieveCrystallisationObligationsRequestParser.scala
@@ -1,36 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package v2.controllers.requestParsers
-
-import api.controllers.requestParsers.RequestParser
-import api.models.domain.Nino
-import v2.controllers.requestParsers.validators.RetrieveCrystallisationObligationsValidator
-import v2.models.request.retrieveCrystallisationObligations.{RetrieveCrystallisationObligationsRawData, RetrieveCrystallisationObligationsRequest}
-import v2.models.request.{ObligationsTaxYear, ObligationsTaxYearHelpers}
-
-import javax.inject.Inject
-
-class RetrieveCrystallisationObligationsRequestParser @Inject()(val validator: RetrieveCrystallisationObligationsValidator)
-    extends RequestParser[RetrieveCrystallisationObligationsRawData, RetrieveCrystallisationObligationsRequest]
-    with ObligationsTaxYearHelpers {
-
-  override protected def requestFor(data: RetrieveCrystallisationObligationsRawData): RetrieveCrystallisationObligationsRequest = {
-    val obligationsTaxYear: ObligationsTaxYear = RawTaxYear(data.taxYear).toObligationsTaxYear
-
-    RetrieveCrystallisationObligationsRequest(Nino(data.nino), obligationsTaxYear)
-  }
-}
