--- app/v2/controllers/requestParsers/validators/RetrieveCrystallisationObligationsValidator.scala
@@ -1,45 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package v2.controllers.requestParsers.validators
-
-import api.controllers.requestParsers.validators.Validator
-import api.controllers.requestParsers.validators.validations._
-import api.models.errors.{MtdError, RuleTaxYearNotSupportedError}
-import v2.models.request.retrieveCrystallisationObligations.RetrieveCrystallisationObligationsRawData
-
-class RetrieveCrystallisationObligationsValidator extends Validator[RetrieveCrystallisationObligationsRawData] {
-
-  private val validationSet = List(parameterFormatValidation, parameterRuleValidation)
-
-  private def parameterFormatValidation: RetrieveCrystallisationObligationsRawData => List[List[MtdError]] = data => {
-    List(
-      NinoValidation.validate(data.nino),
-      data.taxYear.map(TaxYearValidation.validate).getOrElse(Nil)
-    )
-  }
-
-  private def parameterRuleValidation: RetrieveCrystallisationObligationsRawData => List[List[MtdError]] = data => {
-    List(
-      data.taxYear.map(MtdTaxYearValidation.validate(_, RuleTaxYearNotSupportedError)).getOrElse(Nil)
-    )
-  }
-
-  override def validate(data: RetrieveCrystallisationObligationsRawData): List[MtdError] = {
-    run(validationSet, data).distinct
-  }
-
-}
