--- app/v2/models/response/retrieveCrystallisationObligations/des/DesObligation.scala
@@ -1,42 +1,0 @@
-/*
- * Copyright 2023 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package v2.models.response.retrieveCrystallisationObligations.des
-
-import api.models.domain.status.DesStatus
-import play.api.libs.json.{Json, Reads}
-import v2.models.response.retrieveCrystallisationObligations
-
-case class DesObligation(
-    inboundCorrespondenceFromDate: String,
-    inboundCorrespondenceToDate: String,
-    inboundCorrespondenceDueDate: String,
-    status: DesStatus,
-    inboundCorrespondenceDateReceived: Option[String]
-) {
-
-  def toMtd: retrieveCrystallisationObligations.RetrieveCrystallisationObligationsResponse = retrieveCrystallisationObligations.RetrieveCrystallisationObligationsResponse(
-    periodStartDate = inboundCorrespondenceFromDate,
-    periodEndDate = inboundCorrespondenceToDate,
-    dueDate = inboundCorrespondenceDueDate,
-    status = status.toMtd,
-    receivedDate = inboundCorrespondenceDateReceived
-  )
-}
-
-object DesObligation {
-  implicit val reads: Reads[DesObligation] = Json.reads[DesObligation]
-}
