--- conf/v2.routes
@@ -1,1 +1,0 @@
-GET         /:nino/crystallisation                v2.controllers.RetrieveCrystallisationObligationsController.handleRequest(nino: String, taxYear: Option[String])
